Move ByteArrayExtensions to own file
